devtools/client/shared/components/test/mochitest/test_tree_11.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test that when an item in the Tree component is focused by arrow key, the view is scrolled.
-->
<head>
  <meta charset="utf-8">
  <title>Tree component test</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <link rel="stylesheet" href="chrome://devtools/skin/light-theme.css" type="text/css">
  <style>
   .tree {
     height: 30px;
     overflow: auto;
     display: block;
   }

   .tree-node {
     font-size: 10px;
     height: 10px;
   }
  </style>
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
window.onload = async function () {
  try {
    const ReactDOM = browserRequire("devtools/client/shared/vendor/react-dom");
    const { createFactory } = browserRequire("devtools/client/shared/vendor/react");
    const { Simulate } = browserRequire("devtools/client/shared/vendor/react-dom-test-utils");
    const Tree = createFactory(browserRequire("devtools/client/shared/components/VirtualizedTree"));

    TEST_TREE.expanded = new Set("ABCDEFGHIJKLMNO".split(""));

    function renderTree(props) {
      const treeProps = Object.assign({},
        TEST_TREE_INTERFACE,
        {
          itemHeight: 10,
          onFocus: item => renderTree({ focused: item })
        },
        props
      );
      return ReactDOM.render(Tree(treeProps), window.document.body);
    }

    const tree = renderTree({ focused: "K" });

    tree.setState({ scroll: 10 });

    isRenderedTree(document.body.textContent, [
      "A:false",
      "-B:false",
      "--E:false",
      "---K:true",
      "---L:false",
    ], "Should render initial correctly");

    await new Promise(resolve => {
      const treeElem = document.querySelector(".tree");
      treeElem.addEventListener("scroll", function onScroll() {
        dumpn("Got scroll event");
        treeElem.removeEventListener("scroll", onScroll);
        resolve();
      });

      dumpn("Sending ArrowDown key");
      Simulate.keyDown(treeElem, { key: "ArrowDown" });
    });

    dumpn("Forcing re-render");
    await forceRender(tree);

    isRenderedTree(document.body.textContent, [
      "-B:false",
      "--E:false",
      "---K:false",
      "---L:true",
      "--F:false",
    ], "Should have scrolled down one");

  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
};
</script>
</pre>
</body>
</html>
